Add files via upload
The 42 project Born2beroot explores the fundamentals of system administration by inviting us to install and configure a virtual machine. By the end of this project, we will have learned about virtualization, partitioning schemes, LVM, command-line software, protocols such as SSH and the importance of securing a system thanks to sudo, firewalls and password policies.